LayoutTests/platform/ios-sim-deprecated/storage/indexeddb/odd-strings-expected.txt

CONSOLE MESSAGE: line 24: TypeError: null is not an object (evaluating 'indexedDB.deleteDatabase')
Test IndexedDB odd value datatypes

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

Deleting all the databases...
FAIL Unexpected error: TypeError: null is not an object (evaluating 'indexedDB.deleteDatabase')
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
